Improved treatment of d=NaN
In the previous fix, I set a NaN distance to twice the threshold
distance. That's confusing for the user. So now I leave it as NaN, but
treat it internally as greater than the threshold. Thus a NaN
transparently remains a NaN and is interpreted as a very large distance.